Disable CSRF in ErrorsController
When a POST request fails, and ErrorsController is invoked, the CSRF token was checked. In case a redirect is defined for CSRF handling (ie. with a `rescue_from`-clause), this could result in a response code which could be interpreted as success (eg. 302).

We are now disabling CSRF in ErrorsController, as it makes no sense to check the CSRF token when showing an error page.